requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.1
python-multipart>=0.0.18

# NLP
spacy>=3.7.2
trafilatura>=1.8.0
rapidfuzz>=3.9.7
numpy>=1.26.4
scikit-learn>=1.3.2

# AI/LLM
anthropic>=0.49.0

# Backend
firebase-admin>=6.5.0

# HTTP
requests>=2.32.3
httpx>=0.28.0
aiohttp>=3.11.0

# Scraping (Cloudflare Browser Rendering)
# Uses Cloudflare Workers Browser Rendering API via REST

# Export
python-docx>=1.1.0

# Server
gunicorn>=23.0.0

# Optional: OpenAI for dual-engine support
openai>=1.0.0
